dashboard-admin/backend/realtime/requirements.txt
fastapi
uvicorn[standard]
aiokafka
python-dotenv
PyJWT
